organize node-fetching functions in folders
